Update research process approval logic and messages
Added functionality to update the status of a research process to "APPROVED" and check if all related processes are approved. If the current process is the last one, it verifies that all previous processes are approved before updating the research status. The success message displayed to the user is also modified based on the approval status of the last process.